Adding highest quality casino border available
Name assumed to be "bg_border_casino" for the purposes of this border addition. This is from a screenshot sent directly from Xbox One, at 1920x1080 maximum resolution, from Xbox One dataminer and documenter "Blargensnorf". Formatted in format similar to the other borders available. 

To access the casino border on the Xbox One version, you have to donate to the box and get the casino table. Talking to Sans there may or may not also be required.