Corrections are copied to the profiles area if they do not exist, other changes
Code corrections are now stored in "Corrections" in the UndertaleModTool folder.
Profiles are now read, written, and stored in the "UndertaleModTool" folder in AppData
Corrections are copied on load into the profiles area (if they do not yet already exist in those directories).
If the corresponding profile code doesn't exist, then load the regular decompiled output.
Opening the possible objects menu for asset resolution copies the possible objects to the clipboard (easier asset resolution).
GML code changes are only saved to disk if the changes are ultimately successful, otherwise don't save them.
Fix the moving of files from main directory to recovered directory