More Profile Mode Improvements.
When opening a data.win, all the code in the Temp Profile folder is replaced with the main, and when saving, all the main is replaced with the temp. this, to avoid wrong overwriting. Also, exporting profiles will actually overwrite the already exported profiles, and importing too. Fixed the issue where Importing Profiles wouldn't edit the dissasembly code.