additions for Right Click menu of TreeView
Discussed with some people in the discord server, and they encouraged me to make a PR, so uh, this is my attempt at that.

Adds a few things to the right click context menu of the treeview:

New items I added:
<Seperator/>
Move up to {Previous Section}
Move down to {Next Section}
<Seperator/>
Collapse all
Collapse {Current Section}
<Seperator/>
Add: {CurrentSection}

Where Current/Previous/Next Section is replaced with the text of the that section such as "Sounds" "Tile sets" "Sprites", etc etc.