Add multiline regex option for code searching.
Implements #2144. Adds new "Multiline regex" checkbox to Search in code
UI. When the box is checked while doing a regex search, the search is
treated as a multiline regular expression (^ and $ match line start and
end instead of string start and end). The option is only visible when
doing a regex search.